Setup snapshot AMI builds
Setup a convention to build templates in folders named `LATEST`.
Refactor build script to force build any packer template.json that
we find in `*-LATEST` directory.  The idea is to use LATEST as a
convention to build and deploy AMI snapshots.  Directories with a
version number are immutable and will be skipped if the AMI already
exists on the AWS account.