Change to private AMIs (#20)
There are two problems with making our jumpcloud AMIs public:

* Installation of jumpcloud agents requires a "CONNECT_KEY"[1] which
gets saved to a file on the system. That means anyone who provisions an
instance from our AMI will have access to the CONNECT_KEY and we
don't want that.

* Instances provisioned from our AMI will get automatically
associated with our jumpcloud SAS account.  We don't want instances
to be associated with our jumpcloud account unless they are
provisioned by Sage.

We change it to deploy a private AMI that is only share to sage accounts
(sandbox, scicomp, and itsandbox)

[1] https://support.jumpcloud.com/support/s/article/agent-deployment1#linux